requirements.txt
txtsplit
torch
torchaudio
cached_path
tokenizers==0.20.0
transformers==4.45.1
num2words==0.5.12
pykakasi==2.2.1
g2p_en==2.1.0
anyascii==0.3.2
jamo==0.4.1
gruut[de,es,fr]==2.2.3
g2pkk>=0.1.1
librosa==0.9.1
pydub==0.25.1
eng_to_ipa==0.0.2
inflect==7.0.0
unidecode==1.3.7
pypinyin==0.50.0
cn2an==0.5.22
jieba==0.42.1
langid==1.1.6
tqdm
tensorboard==2.16.2
loguru==0.7.2
functions-framework>=3.0.0
google-cloud-storage
cloudevents
